+++ x/inflation/keeper/inflation.go
@@ -1,6 +1,8 @@
 package keeper
 
 import (
+	"strings"
+
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
 	ethermint "github.com/tharsis/ethermint/types"
@@ -94,14 +96,64 @@ func (k Keeper) GetProportions(
 	)
 }
 
+func (k Keeper) isMainnetChainID(ctx sdk.Context) bool {
+	return strings.Contains(ctx.ChainID(), "evmos_9001-")
+}
+
 // BondedRatio the fraction of the staking tokens which are currently bonded
 // It doesn't consider team allocation for inflation
 func (k Keeper) BondedRatio(ctx sdk.Context) sdk.Dec {
 	stakeSupply := k.stakingKeeper.StakingTokenSupply(ctx)
-	if !stakeSupply.IsPositive() || stakeSupply.LTE(teamAlloc) {
+
+	isMainnet := k.isMainnetChainID(ctx)
+
+	if !stakeSupply.IsPositive() || (isMainnet && stakeSupply.LTE(teamAlloc)) {
 		return sdk.ZeroDec()
 	}
 
-	stakeSupply = stakeSupply.Sub(teamAlloc)
+	// don't count team allocation in bonded ratio's stake supple
+	if isMainnet {
+		stakeSupply = stakeSupply.Sub(teamAlloc)
+	}
+
 	return k.stakingKeeper.TotalBondedTokens(ctx).ToDec().QuoInt(stakeSupply)
+}
+
+// GetTotalSupply returns the bank supply of the mintDenom excluding the team
+// allocation in the first year
+func (k Keeper) GetTotalSupply(ctx sdk.Context) sdk.Dec {
+	mintDenom := k.GetParams(ctx).MintDenom
+
+	totalSupply := k.bankKeeper.GetSupply(ctx, mintDenom).Amount.ToDec()
+	teamAllocation := teamAlloc.ToDec()
+
+	// Consider team allocation only on mainnet chain id
+	if k.isMainnetChainID(ctx) {
+		totalSupply = totalSupply.Sub(teamAllocation)
+	}
+
+	return totalSupply
+}
+
+// GetInflationRate returns the inflation rate for the current period.
+func (k Keeper) GetInflationRate(ctx sdk.Context) sdk.Dec {
+	epochMintProvision, _ := k.GetEpochMintProvision(ctx)
+	if epochMintProvision.IsZero() {
+		return sdk.ZeroDec()
+	}
+
+	epp := k.GetEpochsPerPeriod(ctx)
+	if epochMintProvision.IsZero() {
+		return sdk.ZeroDec()
+	}
+
+	epochsPerPeriod := sdk.NewDec(epp)
+
+	totalSupply := k.GetTotalSupply(ctx)
+	if totalSupply.IsZero() {
+		return sdk.ZeroDec()
+	}
+
+	// EpochMintProvision * 365 / totalSupply * 100
+	return epochMintProvision.Mul(epochsPerPeriod).Quo(totalSupply).Mul(sdk.NewDec(100))
 }